refactor(download): extract filename/extension helpers out of RommAPIClient
Move downloadFileExtension and Content-Disposition parsing into standalone
DownloadFilename and ContentDispositionParser types so they can be unit-tested
without the API client.
